# Speed up comparison of monomial orderings