Speed up the comparison of monomial orderings by first checking if the orderings are defined in the exact same way.

I introduce a function `is_obviously_equal` that gets called first in `==` for monomial orderings. Only if it returns `false`, the canonical matrices are compared which is comparatively expensive.
Note that, for example, in the computation of a preimage of a polynomial under a morphism of affine algebras, there are 8 comparisons of monomial orderings performed and 7 of them are 'obviously equal'.
The same should be true for other operations which use the OSCAR <-> Singular interface.

I hope this makes sense and I'm not missing anything.

CC @HechtiDerLachs because I vaguely remember you suffered from the 'slow' comparison of orderings?