@fingolfin 2024-04-26
If @HereAround has any tests where this matters it would be cool to know the timings before/after. But of course this is necessary at all, I am merely curious :-)