@HereAround 2024-04-26
> If @HereAround has any tests where this matters it would be cool to know the timings before/after. But of course this is necessary at all, I am merely curious :-)

At the top of my head, I do not have a good (toric) example yet. Maybe @HechtiDerLachs has such an example?