@HechtiDerLachs 2024-04-26
> I vaguely remember you suffered from the 'slow' comparison of orderings?

I did, thanks for addressing this! I will probably have a look on Monday. If I find any speed improvements, I will post it here.