@HechtiDerLachs 2024-04-26
Before:
```
julia> @time algebraic_lattice(X)
138.240615 seconds (393.46 M allocations: 15.416 GiB, 12.16% gc time)
```
After:
```
julia> @time algebraic_lattice(X)
120.140080 seconds (378.10 M allocations: 15.222 GiB, 12.20% gc time)
```
Looks like a difference. Thanks again!