@joschmitt 2024-04-29
Here are my timings of HechtiDerLachs' example (by the way, I think you could use a new computer), for reference.
I put everything in one function:
```
function hechti()
         k = GF(113)
         kt,t = polynomial_ring(k,:t)
         Ft = fraction_field(kt)

         E = elliptic_curve(Ft, Ft.([0,0,0,(56*t^8 + 56),0]));
         basis_mwl = [[112*t^4 + 112*t^3 + 56*t^2 + 15*t + 1, 100*t^6 + 24*t^5 + 56*t^4 + 13*t^3 + 64*t^2 + 89*t + 31],
                      [31*t^4 + 15*t^2 + 82, 44*t^5 + 14*t^3 + 69*t],
                      [82*t^4 + 13, 37*t^4 + 10],
                      [91*t^4 + 16*t^3 + 25*t^2 + 14*t + 22, 18*t^6 + 55*t^5 + 45*t^4 + 44*t^3 + 110*t^2 + 58*t + 69],
                      [32*t^4 + 72*t^2 + 81, 78*t^6 + 79*t^4 + 58*t^2 + 40],
                      [56*t^4 + 49*t^3 + 85*t^2 + 57*t + 57, 72*t^6 + 25*t^5 + 22*t^4 + 101*t^3 + 104*t^2 + 88*t + 50],
                      [22*t^4 + 87*t^3 + 77*t^2 + 26*t + 22, 44*t^6 + 27*t^5 + 68*t^4 + 98*t^3 + 45*t^2 + 27*t + 69],
                      [112*t^4 + 44*t^3 + 49*t^2 + 44*t + 112, 100*t^6 + 74*t^5 + 49*t^4 + 8*t^3 + 49*t^2 + 74*t + 100]]
         basis_mwl = [E(i) for i in basis_mwl];

         X = elliptic_surface(E, 2, basis_mwl)
         return algebraic_lattice(X)
       end
```
Without this pull request:
```
julia> @btime hechti();
  31.377 s (405598995 allocations: 15.50 GiB)

julia> @btime hechti();
  33.289 s (415381537 allocations: 15.89 GiB)
```
With this pull request:
```
julia> @btime hechti();
  27.040 s (350218521 allocations: 13.77 GiB)

julia> @btime hechti();
  27.006 s (341875932 allocations: 13.42 GiB)
```
I'm in particular impressed (and bit horrified) by the change in allocations.