@HechtiDerLachs 2024-04-29
> I'm in particular impressed (and a bit horrified) by the change in allocations.

I am pointing this out since more than a year now in different occasions. So I'm not particularly surprised by this, but glad for the improvement. 

> by the way, I think you could use a new computer

Admittedly, tests were run on my old one. But working on that machine makes sure you program efficient code.